+++ packages/saltcorn-data/base-plugin/viewtemplates/show.js
@@ -111,7 +111,7 @@ const configuration_workflow = (req) =>
             (f) => f.type && f.type.name === "Bool"
           );
           const stateActions = Object.entries(getState().actions).filter(
-            ([k, v]) => !v.disableInBuilder
+            ([k, v]) => !v.disableInBuilder && !v.disableIf?.()
           );
           const builtInActions = [
             "Delete",