+++ packages/saltcorn-data/db/connect_mobile.ts
@@ -4,6 +4,7 @@ export const getConnectObject = () => {
     sqlite_db_name: "scdb.sqlite",
     default_schema: "public",
     fixed_configuration: {},
+    exposed_configuration: {},
     inherit_configuration: [],
   };
 };