+++ packages/server/routes/view.js
@@ -43,7 +43,10 @@ router.get(
   ["/:viewname", "/:viewname/*slug"],
   error_catcher(async (req, res) => {
     const state = getState();
-    const maintenanceModeEnabled = state.getConfig("maintenance_mode_enabled", false);
+    const maintenanceModeEnabled = state.getConfig(
+      "maintenance_mode_enabled",
+      false
+    );
     const maintenanceModePage = state.getConfig("maintenance_mode_page", "");
 
     if (
@@ -242,7 +245,10 @@ router.post(
   "/:viewname/:route",
   error_catcher(async (req, res, next) => {
     const state = getState();
-    const maintenanceModeEnabled = state.getConfig("maintenance_mode_enabled", false);
+    const maintenanceModeEnabled = state.getConfig(
+      "maintenance_mode_enabled",
+      false
+    );
 
     if (maintenanceModeEnabled && (!req.user || req.user.role_id > 1)) {
       res.status(503).json({ error: "in maintenance mode" });
@@ -287,7 +293,10 @@ router.post(
   ["/:viewname", "/:viewname/*slug"],
   error_catcher(async (req, res, next) => {
     const state = getState();
-    const maintenanceModeEnabled = state.getConfig("maintenance_mode_enabled", false);
+    const maintenanceModeEnabled = state.getConfig(
+      "maintenance_mode_enabled",
+      false
+    );
     if (maintenanceModeEnabled && (!req.user || req.user.role_id > 1)) {
       res.status(503).send("Page Unavailable: in maintenance mode");
       return;