+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Notable changes
 
+## 1.3.1 - In development
+
+* Real time collaboration for Edit views. Form fields will update in real time as they change in the database.
+
+* loop_rows action: repeat an action for a all or some rows in a table, including an option for a random selection of rows.
+
+* train_model_instance action: (re)train a (predictive analytics) model instance
+
+* Visibility toggle option for password fieldviews. Click an eye icon to show password
+
+* Enable/disable push notifications per-role in the Users role table.
+
+* Table triggers (Insert, Update, Delete) and Login and PageLoad can now be limited with an only-if formula.
+
 ## 1.3.0 - Released 2 July 2025
 
 * It is no longer necessary to restart the application when upgrading a plugin to a more recent version